Rename `env` -> `context` everywhere.
This maintains backwards compatibility by passing the old `KUBE_ENV`
environment variable as before, in additino to the new `KUBE_CONTEXT`.

The move makes "context" the consistent method by which a target cluster
is chosen - this is entirely a deploy tooling specific thing as K8s
clusters don't have names. This means that the use of context and the
`KUBE_CLUSTER` envvar are at the discretion of the individual or team
using the tool as a way to differentiate between multiple clusters.