Use containerContext to find a container to run a job
Previous code used the build/dockerfile to find a container matching matching the
one wanted to run a job. However this is not accurate since:
* a dockerfile could be used by different containers
* a container may not have a dockerfile if it's running of a standard image

Using containerContext is a little more accurate since it describes more
precisely the container than just the dockerfile. Also any container with
a dockerfile will anyway require a containercontext.