Fix ktd test / running of commands in containers
In v13 commands in kubetools config are translated to entrypoint in
docker compose. When an entrypoint is present it is always run in
container, with any command being passed to it as argument.

This means that currently running `ktd test` will run any backend start
command present rather than the provided test command.

N.B. This is a breaking change. The only repo I see effected is
https://github.com/EDITD/docs/blob/da424553457ebe5a263c36e58838ddd737349c4d/manifests/es_knn_exporter/kubetools.yml