Add support for job time to live
Failure to remove resources used by completed jobs/pods can lead to
resource exhaustion in the Kubernetes cluster. This commit adds support
for setting a ttlSecondsAfterFinished parameter in the Kubernetes Job
specification. This parameter specifies the amount of time (in seconds)
for which the resources used by completed job/pod will persist after it
which it will be garbage collected by the Kube control manager.

The newly added parameter is optional and only set if it passed in
through the `config` argument to the `make_job_spec` method.